Meet suggestions implemented

+++ app/src/main/java/com/phad/chatapp/models/AttendanceEvent.kt
@@ -30,13 +30,13 @@ data class AttendanceEvent(
     val eventTime: String = "", // Time range in "HH:MM AM/PM - HH:MM AM/PM" format
 
     @PropertyName("hours")
-    val hours: Int = 0, // Volunteer hours/tokens awarded for event participation
+    val hours: Double = 0.0, // Volunteer hours/tokens awarded for event participation
 
     @PropertyName("mandatory")
     val isMandatory: Boolean = false, // Whether attendance is mandatory
 
     @PropertyName("negativeHours")
-    val negativeHours: Int = 0, // Hours to deduct for absentees when mandatory
+    val negativeHours: Double = 0.0, // Hours to deduct for absentees when mandatory
 
     @PropertyName("location")
     val location: String = "", // Location where the event will take place
@@ -72,9 +72,9 @@ data class AttendanceEvent(
         id = "",
         eventDate = "",
         eventTime = "",
-        hours = 0,
+        hours = 0.0,
         isMandatory = false,
-        negativeHours = 0,
+        negativeHours = 0.0,
         location = "",
         description = "",
         createdBy = "",
